fix(reactions): fourth-pass audit — all remaining issues resolved
## Fixes

### H1 — Advisory lock int64 cast (reaction_service.go:106)
`pg_advisory_xact_lock` expects a bigint. Go's `int` is platform-dependent
(32-bit on 32-bit systems). Explicitly cast `messageID` to `int64` before
passing to avoid silent overflow on any platform where `int` is narrower
than 64 bits.

### M1 — Full lifecycle test
`TestReactionLifecycle_AddGetRemoveGet`: validates the complete round-trip —
AddReaction → GetReactions (count=1) → RemoveReaction → GetReactions (count=0).
This was the one test that exercises the feature as a user would experience it.

### M2 — Concurrent cap enforcement test
`TestAddReaction_ConcurrentCapEnforcement`: fires 11 goroutines simultaneously,
each attempting to add a distinct emoji to the same message. Verifies that the
`pg_advisory_xact_lock` serialises the cap-check + insert correctly: exactly
10 succeed (the cap) and exactly 1 is rejected with 409 Conflict.

### M3 — RemoveReaction broadcast test
`TestRemoveReaction_BroadcastsToOtherParticipant`: mirrors the existing
AddReaction broadcast test for the removal path. Verifies that `reaction_removed`
is delivered to user2 (the non-actor participant) within 500 ms and that user1
(the actor) is not broadcast to.

## Test count
- handlers/reactions_test.go: 28 tests (was 25) — all PASS
- services/reaction_service_test.go: 38 tests — all PASS
- Total: 66 reaction tests passing

+++ backend/internal/services/reaction_service.go
@@ -103,7 +103,7 @@ func (s *ReactionService) AddReaction(ctx context.Context, messageID, userID int
 	}
 	defer tx.Rollback(ctx)
 
-	if _, err := tx.Exec(ctx, `SELECT pg_advisory_xact_lock($1::bigint)`, messageID); err != nil {
+	if _, err := tx.Exec(ctx, `SELECT pg_advisory_xact_lock($1::bigint)`, int64(messageID)); err != nil {
 		return nil, fmt.Errorf("acquire emoji cap lock: %w", err)
 	}
